fix: spice_level SMALLINT to INTEGER (Hibernate validate mismatch on prod)
First 0.1.8 deploy failed with schema validation error because
V4 declared menu_item.spice_level as SMALLINT but the JPA entity
maps Integer to Types#INTEGER (PG int4). Tests passed because the
reused Testcontainers PG happened to already have int4 from a
prior V4 variant; prod was fresh and got the literal SMALLINT.

* V4: declare spice_level as INT (int4) to match the entity
* V5: ALTER menu_item.spice_level TYPE INTEGER (one-shot repair
  for prod where V4 already ran with SMALLINT)

Bumps 0.1.8 -> 0.1.9 (the 0.1.8 release never came up).

+++ src/main/resources/db/migration/V4__product_metadata.sql
@@ -15,7 +15,7 @@ ALTER TABLE product
     ADD COLUMN weight_grams  INT;
 
 ALTER TABLE menu_item
-    ADD COLUMN spice_level   SMALLINT;
+    ADD COLUMN spice_level   INT;
 
 ALTER TABLE product
     ADD CONSTRAINT ck_product_weight_grams
